requirements.txt
Django>=1.11
django-filter==2.0.0
django-sendsms==0.3.1
djangorestframework>=3.8.0
djangorestframework-jwt==1.11.0
drfaddons>=0.1.0
